+++ spec/models/document_paranoid.rb
@@ -1,9 +1,0 @@
-class DocumentParanoid
-  include Mongoid::Document
-  # slug, then paranoia
-  include Mongoid::Slug
-  include Mongoid::Paranoia
-
-  field :title
-  slug :title
-end