+++ spec/models/paranoid_document.rb
@@ -1,8 +1,0 @@
-class ParanoidDocument
-  include Mongoid::Document
-  include Mongoid::Paranoia
-  include Mongoid::Slug
-
-  field :title
-  slug :title
-end