+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 5.3.4 (Next)
 
-* Your contribution here.
+* [#251](https://github.com/mongoid/mongoid-slug/pull/251): Use Mongoid::Config.models to get the list of all available models in mongoid_slug.rake - [@syzspectroom](https://github.com/syzspectroom).
 
 ## 5.3.3 (2017/04/06)