+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-## 5.3.4 (Next)
+## 6.0.0 (Next)
 
-* Your contribution here.
+* [#252](https://github.com/mongoid/mongoid-slug/pull/252): Compatibility with Mongoid 7.0.0+ - [@kailan](https://github.com/kailan).
 
 ## 5.3.3 (2017/04/06)